Add init script
